fix(cli): eliminate image-mode flashing and move frame rendering to display.ts
The -i image mode flashed on animated apps while the ANSI fallback was
smooth — three causes in how the native graphics protocols handle
inline images:

- Frames carried transparency (apps with SetAlphaEnable): iTerm2/Kitty
  honor the alpha channel, letting the terminal theme color show
  through and flash as alpha regions move. Frames are now flattened
  onto opaque black before PNG encoding.
- Synchronized-output mode (DEC 2026) now opens BEFORE invoking
  terminal-image: its Kitty path writes the image escape directly to
  stdout, so the previous wrapper closed around nothing. With 2026 in
  effect the terminal swaps old/new frames atomically instead of
  blanking while the payload decodes (unsupported terminals ignore it).
- The iTerm2 header gets doNotMoveCursor=1 (ansi-escapes does not
  expose it) and the cursor is hidden for the whole run, so repaints
  can never scroll or flicker at the image edge.

All frame-rendering machinery also moved from index.ts to display.ts
(enable/disableFrameOutput, newest-frame coalescing, backpressure
handling, image/ascii renderers, frameToPng). The display module no
longer reads the CLI program object — enableFrameOutput takes a
DisplayMode built from the parsed options.
